+++ bin/run-repl-quick.sh
@@ -13,7 +13,7 @@ set -e
 
 SCRIPT_DIR=$(dirname $0)
 
-STREAMLY_VERSION=0.8.2
+STREAMLY_VERSION=0.8.3
 BENCH_REPORT_DIR=benchmark/bench-report
 source $SCRIPT_DIR/targets.sh